% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_string}
\alias{add_string}
\title{Concatenate Strings Using the \code{+} Operator}
\usage{
add_string(x)
}
\arguments{
\item{x}{An expression using the \code{+} operator for string concatenation.}
}
\value{
A character string resulting from concatenating the provided components.
}
\description{
Provides syntactic sugar for string concatenation, allowing the use of \code{+} as a shorthand for \code{paste0()}.
}
\details{
This function evaluates an expression in a temporary environment where \code{+} is redefined
as \code{paste0()}. It enables writing cleaner, more readable string concatenations.
}
\examples{
add_string("Hello, " + "world" + "!")
# Returns "Hello, world!"

}
