% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRIDGE21-package.R
\name{BRIDGE21}
\alias{BRIDGE21}
\title{BRIDGE21: Building Reports and Insights for Down syndrome Guidance and Empowerment}
\description{
Generates participant feedback reports and insights for Down syndrome
cohorts to help automate the process of getting the data back into the hands
of the participants and their families.
}
\section{BRIDGE21 functions}{


\code{\link{generate_report}}
}

