% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blood_pressure.R
\name{classify_blood_pressure}
\alias{classify_blood_pressure}
\title{classify_blood_pressure}
\usage{
classify_blood_pressure(value, type = c("systolic", "diastolic"))
}
\value{
A risk category based on the value and type of blood pressure measure.
}
\description{
Classifies systolic or diastolic blood pressure and returns the
level of associated risk.
}
\details{
Classifies systolic or diastolic blood pressure and returns the
level of associated risk.
}
\keyword{internal}
