% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blood_pressure.R
\name{create_blood_pressure_table}
\alias{create_blood_pressure_table}
\title{create_blood_pressure_table}
\usage{
create_blood_pressure_table()
}
\value{
A table containing blood pressure and risk categories
}
\description{
Creates a table containing systolic and diastolic blood pressure
values and their respective risk categories.
}
\details{
Creates a table containing systolic and diastolic blood pressure
values and their respective risk categories.
}
\keyword{internal}
