% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dxa.R
\name{generate_body_fat_plot}
\alias{generate_body_fat_plot}
\title{generate_body_fat_plot}
\usage{
generate_body_fat_plot(total_fat, subject_sex, subject_birthdate, fittest_date)
}
\arguments{
\item{total_fat}{Participant's body fat percent}

\item{subject_sex}{Participant's sex (1 = male, 0 = female)}

\item{subject_birthdate}{Date of birth (YYYY-MM-DD)}

\item{fittest_date}{The date of a DXA scan in a \%m/\%d/\%Y format}
}
\value{
A ggplot object
}
\description{
Generates a bar plot of the participant's body fat percent compared to norms by age and sex.
}
