% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blood_pressure.R
\name{generate_bp_plot}
\alias{generate_bp_plot}
\title{generate_bp_plot}
\usage{
generate_bp_plot(bpsys, bpdia)
}
\arguments{
\item{bpsys}{Systolic Blood Pressure}

\item{bpdia}{Diastolic Blood Pressure}
}
\value{
A ggplot2 object containing the plotted blood pressure values
}
\description{
Generate a plot containing the systolic and diastolic blood pressure values
}
\details{
Generate a plot containing the systolic and diastolic blood pressure values
}
\seealso{
\code{\link[magick]{editing}}
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_raster}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{annotation_custom}}, \code{\link[ggplot2]{annotate}}, \code{\link[ggplot2]{geom_label}}, \code{\link[ggplot2]{ggtheme}}
\code{\link[grid]{grid.raster}}, \code{\link[grid]{patterns}}
}
