% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{generate_report}
\usage{
generate_report(
  ...,
  id,
  datafile,
  outputdir,
  acceldir = NULL,
  example_report = FALSE
)
}
\arguments{
\item{...}{A vector of unquoted or quoted reports to generate from .allReports}

\item{id}{The patient id as represented in the study database}

\item{datafile}{The path name of the file containing the data as a .csv file}

\item{outputdir}{The path name of the directory containing the participant feedback reports}

\item{acceldir}{The path name of the directory containing the raw .gt3x accelerometer files, Default = NULL}

\item{example_report}{Used with outputdir to generate a sample report using example data, Default = FALSE}
}
\value{
The location of the participant feedback report
}
\description{
Generates a Quarto pdf report containing participant feedback from the study visit
}
\details{
Generates a Quarto pdf report containing participant feedback from the study visit
}
\seealso{
\code{\link[readr]{read_delim}}
\code{\link[kuadrc.xnat]{get_projects}}, \code{\link[kuadrc.xnat]{get_experiments}}, \code{\link[kuadrc.xnat]{get_scans}}, \code{\link[kuadrc.xnat]{xnat_download}}, \code{\link[kuadrc.xnat]{convert_to_nifti}}, \code{\link[kuadrc.xnat]{save_nifti_image}}
\code{\link[yaml]{write_yaml}}
\code{\link[quarto]{quarto_render}}
}
