% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bmi_weight.R
\name{generate_weight_bar_plot}
\alias{generate_weight_bar_plot}
\title{generate_weight_bar_plot}
\usage{
generate_weight_bar_plot(wt, ht, metric = TRUE)
}
\arguments{
\item{wt}{Research participant's weight in kilograms or pounds}

\item{ht}{Research participant's height in centimeters or inches}

\item{metric}{A parameter to inform the function of using the metric units of kilograms and centimeters, Default: TRUE}
}
\value{
A bar plot of the participant's weight and its comparison to the weight at a healthy body mass index.
}
\description{
Generates a bar plot of the participant's weight and its
comparison to the weight at a healthy body mass index.
}
\details{
Generates a bar plot of the participant's weight and its
comparison to the weight at a healthy body mass index.
}
\seealso{
\code{\link[ggplot2]{geom_segment}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{annotate}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_raster}}, \code{\link[ggplot2]{geom_abline}}, \code{\link[ggplot2]{scale_continuous}}, \code{\link[ggplot2]{labs}}, \code{\link[ggplot2]{coord_fixed}}, \code{\link[ggplot2]{theme}}, \code{\link[ggplot2]{element}}, \code{\link[ggplot2]{reexports}}
}
