% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anthropometrics.R
\name{plot_body_mass_index}
\alias{plot_body_mass_index}
\title{plot_body_mass_index}
\usage{
plot_body_mass_index(wt, ht, metric = TRUE)
}
\arguments{
\item{wt}{Research participant's weight in kilograms or pounds}

\item{ht}{Research participant's height in centimeters or inches}

\item{metric}{A parameter to inform the function of using the metric units of kilograms and centimeters, Default: TRUE}
}
\value{
An arch plot of the participants body mass index and its associated risk level.
}
\description{
Generates an arch plot of the participant's body mass index and
its associated risk level for cardiometabolic conditions.
}
\details{
Generates an arch plot of the participant's body mass index and
its associated risk level for cardiometabolic conditions.
}
\seealso{
\code{\link[ggplot2]{geom_polygon}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_label}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_raster}}, \code{\link[ggplot2]{annotate}}, \code{\link[ggplot2]{scale_continuous}}, \code{\link[ggplot2]{coord_polar}}, \code{\link[ggplot2]{ggtheme}}
}
