% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series_by_date.R
\name{plot_time_series_by_date}
\alias{plot_time_series_by_date}
\title{plot_time_series_by_date}
\usage{
plot_time_series_by_date(IMP, summary_data)
}
\arguments{
\item{IMP}{The the path to the file containing the IMP object from GGIR Part 2}

\item{summary_data}{A data frame with date and min/day of sleep, inactivity, light, and moderate-to-vigorous physical activity}
}
\value{
A ggplot2 object
}
\description{
Plot a line graph of the time series acceleration by date
}
\details{
Plot a line graph of the time series acceleration by date
}
\seealso{
\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{mutate}}
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_path}}, \code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{facet_wrap}}, \code{\link[ggplot2]{vars}}, \code{\link[ggplot2]{ggtheme}}, \code{\link[ggplot2]{theme}}, \code{\link[ggplot2]{element}}
}
