% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blood_pressure.R
\name{round_rectangle}
\alias{round_rectangle}
\title{round_rectangle}
\usage{
round_rectangle(x, y, width, height, radius, fill)
}
\arguments{
\item{x, y, width, height}{The location and size of the rectangle.}

\item{fill}{Fill color of the rectangle}
}
\value{
An annotation_custom ggplot2 layer
}
\description{
A convenience function to plot the rounded rectangles in generate_bp_plot
}
\details{
A convenience function to plot the rounded rectangles in generate_bp_plot
}
\seealso{
\code{\link[ggplot2]{annotation_custom}}
\code{\link[grid]{roundrect}}, \code{\link[grid]{gpar}}
}
\keyword{internal}
