% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ggir.R
\name{run_ggir}
\alias{run_ggir}
\title{run_ggir}
\usage{
run_ggir(id, acceldir, outputdir)
}
\arguments{
\item{id}{The patient id as represented in the study database}

\item{acceldir}{The path name of the directory containing the raw .gt3x accelerometer files}

\item{outputdir}{The path name of the directory containing the participant feedback reports}
}
\value{
The path name to the results of the processed accelerometer data
}
\description{
Processes the ActiGraph accelerometer data with the GGIR R Package
}
\details{
Processes the ActiGraph accelerometer data with the GGIR R Package
}
\seealso{
\code{\link[GGIR]{GGIR}}
}
