% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_xnat_images.R
\name{save_xnat_images}
\alias{save_xnat_images}
\title{save_xnat_images}
\usage{
save_xnat_images(imagedir, data, id)
}
\arguments{
\item{imagedir}{The directory where the MRI images should be stored}

\item{data}{The data set containing the patient's information}

\item{id}{The patient id as represented in the study database}
}
\description{
Retrieves MRI images from the XNAT API and saves as png files
}
\details{
Retrieves MRI images from the XNAT API and saves as png files
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[kuadrc.xnat]{get_projects}}, \code{\link[kuadrc.xnat]{get_subjects}}, \code{\link[kuadrc.xnat]{get_experiments}}, \code{\link[kuadrc.xnat]{get_scans}}, \code{\link[kuadrc.xnat]{xnat_download}}, \code{\link[kuadrc.xnat]{convert_to_nifti}}, \code{\link[kuadrc.xnat]{save_nifti_image}}
}
