% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_xnat_images.R
\name{update_xnat_images}
\alias{update_xnat_images}
\title{update_xnat_images}
\usage{
update_xnat_images(id, outputdir, ...)
}
\arguments{
\item{id}{The patient id as represented in the study database}

\item{outputdir}{The path name of the directory containing the participant feedback reports}

\item{...}{Additional arguments passed to shiny::shinyApp}
}
\value{
A Shiny app launched in the local browser to adjust the MRI images with a save button
}
\description{
Launches a Shiny UI to select a slice of the MRI image in the case that the disclosure
produces a blurry or distorted image
}
\details{
Launches a Shiny UI to select a slice of the MRI image in the case that the disclosure
produces a blurry or distorted image
}
\seealso{
\code{\link[shiny]{fluidPage}}, \code{\link[shiny]{titlePanel}}, \code{\link[shiny]{sidebarLayout}}, \code{\link[shiny]{reexports}}, \code{\link[shiny]{radioButtons}}, \code{\link[shiny]{htmlOutput}}, \code{\link[shiny]{numericInput}}, \code{\link[shiny]{actionButton}}, \code{\link[shiny]{plotOutput}}, \code{\link[shiny]{textOutput}}, \code{\link[shiny]{renderUI}}, \code{\link[shiny]{renderPlot}}, \code{\link[shiny]{req}}, \code{\link[shiny]{observeEvent}}, \code{\link[shiny]{showNotification}}, \code{\link[shiny]{shinyApp}}
}
