% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRIDGE21-package.R
\docType{package}
\name{BRIDGE21-package}
\alias{BRIDGE21-package}
\title{BRIDGE21: Building Reports and Insights for Down syndrome Guidance and Empowerment}
\description{
Generates participant feedback reports and insights for Down syndrome cohorts to help automate the process of getting the data back into the hands of the participants and their families.
}
\author{
\strong{Maintainer}: Brian C. Helsel \email{bhelsel@kumc.edu}

Other contributors:
\itemize{
  \item Julianne G. Clina \email{jclina@kumc.edu} [contributor]
  \item University of Kansas [copyright holder]
}

}
\keyword{internal}
